refactor: lift `invalidate_all_tokens` out of `legacy` module
Since it now uses generation IDs, it is generic over the account type.
